move sign out button to the corner, show on all error pages
