+++ app/components/ErrorPage.tsx
@@ -10,6 +10,10 @@ import { Link } from 'react-router-dom'
 
 import { Error12Icon, PrevArrow12Icon } from '@oxide/design-system/icons/react'
 
+import { useApiMutation } from '~/api/client'
+import { navToLogin } from '~/api/nav-to-login'
+import { Button } from '~/ui/lib/Button'
+
 const GradientBackground = () => (
   <div
     // negative z-index avoids covering MSW warning banner
@@ -27,7 +31,7 @@ export function ErrorPage({ children }: Props) {
   return (
     <div className="flex w-full justify-center">
       <GradientBackground />
-      <div className="relative w-full">
+      <div className="relative flex w-full justify-between">
         <Link
           to="/"
           className="flex items-center p-6 text-mono-sm text-secondary hover:text-default"
@@ -35,6 +39,7 @@ export function ErrorPage({ children }: Props) {
           <PrevArrow12Icon title="Select" className="mr-2 text-tertiary" />
           Back to console
         </Link>
+        <SignOutButton className="mr-6 mt-4" />
       </div>
       <div className="absolute left-1/2 top-1/2 flex w-96 -translate-x-1/2 -translate-y-1/2 flex-col items-center justify-center space-y-4 rounded-lg border p-8 !bg-raise border-secondary elevation-3">
         <div className="my-2 flex h-12 w-12 items-center justify-center">
@@ -56,5 +61,21 @@ export function NotFound() {
         The page you are looking for doesn&apos;t exist or you may not have access to it.
       </p>
     </ErrorPage>
+  )
+}
+
+export function SignOutButton({ className }: { className?: string }) {
+  const logout = useApiMutation('logout', {
+    onSuccess: () => navToLogin({ includeCurrent: false }),
+  })
+  return (
+    <Button
+      onClick={() => logout.mutate({})}
+      className={className}
+      size="sm"
+      variant="ghost"
+    >
+      Sign out
+    </Button>
   )
 }
